Fix transitions for screens component
